Updated adapter location
