--- feature-core/src/main/java/movie/core/CompositeException.kt
@@ -1,12 +1,0 @@
-package movie.core
-
-class CompositeException(
-    vararg causes: Throwable
-) : RuntimeException() {
-
-    init {
-        for (cause in causes)
-            addSuppressed(cause)
-    }
-
-}
