+++ feature-core/src/main/java/movie/core/EventPreviewFeatureStoring.kt
@@ -1,5 +1,6 @@
 package movie.core
 
+import movie.core.adapter.asStored
 import movie.core.db.dao.MovieDao
 import movie.core.db.dao.MovieMediaDao
 import movie.core.db.dao.MoviePreviewDao
