+++ feature-core/src/main/java/movie/core/Typealias.kt
